Factor out `file`
