+++ async-openai/src/types/mod.rs
@@ -5,6 +5,7 @@ mod assistant_file;
 mod assistant_impls;
 mod audio;
 mod common;
+mod file;
 mod image;
 mod message;
 mod message_file;
@@ -17,6 +18,7 @@ pub use assistant::*;
 pub use assistant_file::*;
 pub use audio::*;
 pub use common::*;
+pub use file::*;
 pub use image::*;
 pub use message::*;
 pub use message_file::*;
